# lib/api-zod/src/generated/types/acousticMatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AudioPayload } from "./audioPayload";

export interface AcousticMatch {
  phrase: string;
  language: string;
  languageCode: string;
  meaning: string;
  notes: string;
  /** 0-1 acoustic similarity from MFCC+DTW */
  similarity: number;
  dtwDistance: number;
  audio: AudioPayload;
}