# lib/api-zod/src/generated/types/audioPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Encoded audio + numeric features extracted server-side from raw TTS PCM
 */
export interface AudioPayload {
  /** WAV-encoded audio for browser playback */
  wavBase64: string;
  sampleRate: number;
  durationMs: number;
  /** Downsampled waveform peaks for visualization */
  waveform: number[];
  /** Mel-spectrogram frames (rows are frames, columns are mel bins) */
  melSpectrogram: number[][];
  melMin: number;
  melMax: number;
}