# lib/api-zod/src/generated/types/compareRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompareRequest {
  phrase1: string;
  language1?: string;
  phrase2: string;
  language2?: string;
}