# lib/api-zod/src/generated/types/compareResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AudioPayload } from "./audioPayload";

export interface CompareResponse {
  phrase1: string;
  phrase2: string;
  audio1: AudioPayload;
  audio2: AudioPayload;
  similarity: number;
  dtwDistance: number;
  verdict: string;
}