# lib/api-zod/src/generated/types/discoverResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AcousticMatch } from "./acousticMatch";
import type { AudioPayload } from "./audioPayload";

export interface DiscoverResponse {
  sourcePhrase: string;
  sourceLanguage: string;
  sourceLanguageName: string;
  sourceMeaning: string;
  sourceAudio: AudioPayload;
  matches: AcousticMatch[];
  candidatesEvaluated: number;
  elapsedMs: number;
}