# lib/api-zod/src/generated/types/savedPair.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedPair {
  id: number;
  sourcePhrase: string;
  sourceLanguage: string;
  sourceMeaning: string;
  matchPhrase: string;
  matchLanguage: string;
  matchMeaning: string;
  similarity: number;
  notes?: string;
  createdAt: string;
}