# lib/api-zod/src/generated/types/ttsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AudioPayload } from "./audioPayload";

export interface TtsResponse {
  audio: AudioPayload;
}